# src/main/resources/api/openapi.yaml
openapi: 3.0.0
info:
  title: 아무말 대잔치 커뮤니티 API
  version: 0.1.0

servers:
  - url: /api/v1

paths:
  /auth:
    post:
      summary: 로그인
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
              required:
                - email
                - password
            examples:
              basic:
                value:
                  email: "exemple@exemple.com"
                  password: "Password2@"
      responses:
        '200':
          description: 로그인 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: login_success
                  data:
                    type: object
                    properties:
                      session_id:
                        type: string
                      user_id:
                        type: string
                      nickname:
                        type: string
                      email:
                        type: string
                      profile_image:
                        type: string
                        nullable: true
              examples:
                success:
                  value:
                    message: login_success
                    data:
                      session_id: s123
                      user_id: u123
                      nickname: john
                      email: example@example.com
                      profile_image: https://image.kr/img.jpg
        '401':
          description: 인증 실패
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    type: object
                    nullable: true
              examples:
                invalid:
                  value:
                    message: invalid_email_or_password
                    data: null
    delete:
      summary: 로그아웃
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                session_id:
                  type: string
              required:
                - session_id
            examples:
              basic:
                value:
                  session_id: s123
      responses:
        '204':
          description: 로그아웃 성공

  /users:
    post:
      summary: 회원가입
      tags:
        - users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                nickname:
                  type: string
                profile_image:
                  type: string
                  nullable: true
              required:
                - email
                - password
                - nickname
            examples:
              basic:
                value:
                  email: test@startupcode.kr
                  password: test1234
                  nickname: startup
                  profile_image: https://image.kr/img.jpg
      responses:
        '201':
          description: 회원가입 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: signup_success
                  data:
                    type: object
                    nullable: true
              examples:
                success:
                  value:
                    message: signup_success
                    data: null
        '400':
          description: 유효성 오류
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: error
                  error_code:
                    type: string
                    example: VALIDATION_ERROR
                  details:
                    type: array
                    items:
                      type: object
                      properties:
                        field:
                          type: string
                        reason:
                          type: string
        '409':
          description: 중복 이메일
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: error
                  error_code:
                    type: string
                    example: DUPLICATE_EMAIL

  /users/{userId}:
    patch:
      summary: 회원정보수정
      tags:
        - users
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nickname:
                  type: string
                profile_image:
                  type: string
                  nullable: true
            examples:
              basic:
                value:
                  nickname: new_startup
                  profile_image: https://image.kr/new.jpg
      responses:
        '200':
          description: 프로필 수정 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: user_profile_update_success
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                      nickname:
                        type: string
                      profile_image:
                        type: string
                        nullable: true
              examples:
                success:
                  value:
                    message: user_profile_update_success
                    data:
                      id: u123
                      nickname: new_startup
                      profile_image: https://image.kr/new.jpg
    delete:
      summary: 회원탈퇴
      tags:
        - users
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                session_id:
                  type: string
              required:
                - session_id
            examples:
              basic:
                value:
                  session_id: s123
      responses:
        '204':
          description: 회원탈퇴 성공

  /posts:
    get:
      summary: 게시글 목록 불러오기
      tags:
        - posts
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
        - name: cursor
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: 목록 조회 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: get_posts_success
                  data:
                    type: object
                    properties:
                      posts:
                        type: array
                        items:
                          type: object
                      next_cursor:
                        type: string
                        nullable: true
              examples:
                success:
                  value:
                    message: get_posts_success
                    data:
                      posts:
                        - id: p123
                          title: 게시글 제목
                          image: url1
                          created_at: "2025-09-20 10:00:00"
                          view_count: 150
                          like_count: 25
                          comment_count: 47
                          author:
                            id: u123
                            name: 작성자명
                            profile_image: profile_url
                      next_cursor: p124
    post:
      summary: 게시글 작성
      tags:
        - posts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                session_id:
                  type: string
                title:
                  type: string
                image:
                  type: string
                  nullable: true
                content:
                  type: string
              required:
                - session_id
                - title
                - content
            examples:
              basic:
                value:
                  session_id: s123
                  title: 게시글 제목
                  image: image_url
                  content: 본문 내용
      responses:
        '201':
          description: 게시글 생성됨
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      post_id:
                        type: string
              examples:
                success:
                  value:
                    message: 게시글 생성됨
                    data:
                      post_id: p123

  /posts/{postId}:
    get:
      summary: 게시글 상세조회
      tags:
        - posts
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: get_post_success
                  data:
                    type: object
                    properties:
                      post:
                        type: object
                      comments:
                        type: array
                        items:
                          type: object
              examples:
                success:
                  value:
                    message: get_post_success
                    data:
                      post:
                        id: p123
                        title: 게시글 제목
                        content: 본문 내용
                        image: url1
                        created_at: "2025-09-20 10:00:00"
                        view_count: 150
                        like_count: 25
                        comment_count: 47
                        author:
                          id: u123
                          name: 작성자명
                          profile_image: profile_url
                      comments:
                        - id: c123
                          content: 댓글 내용
                          created_at: "2025-09-20 11:00:00"
                          author:
                            id: u123
                            name: 댓글작성자
                            profile_image: profile_url
    patch:
      summary: 게시글 수정
      tags:
        - posts
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                session_id:
                  type: string
                title:
                  type: string
                image:
                  type: string
                  nullable: true
                content:
                  type: string
            examples:
              basic:
                value:
                  session_id: s123
                  title: 수정 제목
                  image: url1
                  content: 수정 내용
      responses:
        '200':
          description: 게시글 수정 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      post_id:
                        type: string
              examples:
                success:
                  value:
                    message: 게시글 수정 성공
                    data:
                      post_id: p123
        '403':
          description: 제어 권한 없음
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    type: object
                    nullable: true
              examples:
                forbidden:
                  value:
                    message: 제어 권한 없음
                    data: null
    delete:
      summary: 게시글 삭제
      tags:
        - posts
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                session_id:
                  type: string
              required:
                - session_id
            examples:
              basic:
                value:
                  session_id: s123
      responses:
        '204':
          description: 게시글 삭제됨
        '403':
          description: 제어 권한 없음
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    type: object
                    nullable: true
              examples:
                forbidden:
                  value:
                    message: 제어 권한 없음
                    data: null

  /posts/{postId}/comments:
    post:
      summary: 댓글 작성
      tags:
        - comments
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                session_id:
                  type: string
                content:
                  type: string
              required:
                - session_id
                - content
            examples:
              basic:
                value:
                  session_id: s123
                  content: 댓글 내용
      responses:
        '201':
          description: 댓글 작성됨
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      comment_id:
                        type: string
              examples:
                success:
                  value:
                    message: 댓글 작성됨
                    data:
                      comment_id: c123

  /posts/{postId}/comments/{commentId}:
    patch:
      summary: 댓글 수정
      tags:
        - comments
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
        - name: commentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                session_id:
                  type: string
                content:
                  type: string
              required:
                - session_id
                - content
            examples:
              basic:
                value:
                  session_id: s123
                  content: 댓글 내용
      responses:
        '200':
          description: 댓글 수정 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      comment_id:
                        type: string
                      content:
                        type: string
              examples:
                success:
                  value:
                    message: 댓글 수정 성공
                    data:
                      comment_id: c123
                      content: 댓글 내용
        '403':
          description: 제어 권한 없음
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    type: object
                    nullable: true
              examples:
                forbidden:
                  value:
                    message: 제어 권한 없음
                    data: null
    delete:
      summary: 댓글 삭제
      tags:
        - comments
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
        - name: commentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                session_id:
                  type: string
              required:
                - session_id
            examples:
              basic:
                value:
                  session_id: s123
      responses:
        '204':
          description: 댓글 삭제됨
        '403':
          description: 제어 권한 없음
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    type: object
                    nullable: true
              examples:
                forbidden:
                  value:
                    message: 제어 권한 없음
                    data: null

  /posts/{postId}/like:
    patch:
      summary: 게시글에 좋아요
      tags:
        - posts
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                session_id:
                  type: string
              required:
                - session_id
            examples:
              basic:
                value:
                  session_id: s123
      responses:
        '200':
          description: 좋아요 토글 결과
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 좋아요 토글됨
                  data:
                    type: object
                    properties:
                      liked:
                        type: boolean
                      like_count:
                        type: integer
              examples:
                success:
                  value:
                    message: 좋아요 토글됨
                    data:
                      liked: true
                      like_count: 23

  /health:
    get:
      summary: 헬스체크
      tags:
        - system
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

tags:
  - name: auth
  - name: users
  - name: posts
  - name: comments
  - name: system